Made Prototype Yarn Options shorter, added art, added replay button
Moved UI elements a bit to account for text length as well as new placement around new art! Also changed some longer options to just the first sentence, and then took the remaining sentences and added them to the top of the corresponding node. Also added a temporary "replay" button that appears at the end of a dialogue.