Created the DialogueManager and CharacterManager
Created the DialogueManager which will host the functionality for the YarnCommands, such that the YarnCommands file can exist purely as a container class that creates the commands for Yarn (as it should be). Also created the CharacterManager class, which will be attached to all characters in a scene, and contains the following info: their dialogue print color, an array of expressions they could cycle between (if in car), an array of screen positions their speech bubble could appear (if in car).

Within DialogueManager, coded a "setSpeaker" method, which takes in a character's name, parses through a public array of character gameobjects, sets the text bubble to that character's color, and moves the bubble to them. If it is not an exploration scene, it will move it to the first position in that character's list of possible positions. If it is in exploration, currently it moves the bubble to the same position as the GameObject. (This will likely need to go through localPositions or w/e when the map gets bigger.)

All testing for this was done in the sample scene.