Finished Options move functionality, added comments, hover Options text
The options bubbles now move whenever SetBubblePosition is called in accordance to Daniella. Also, updated TextDisplayHandler -> OptionsTextDisplayHandler, which now implements pointer interfaces such that box colliders don't have to be used, as adjusting the colliders to change alongside the dynamic canvas elements was a pain. So now, the Options buttons' text will change color on entry/exit/down, like they did in the prototype, but it's ~dynamic.