Finished setting up Chapter 1 Scene & prefabs, added AlphaPlaytest yarn
Daniella and Mish have arbitrary speech bubble colors attached to the prefabs now. Note that the prefabs are organized according to who is driving and is the passenger, as they'll need slightly different bubble positions depending on where they're sitting. But this was we can have prefabs with the flipped sprites and correct bubble positions ready to throw into scenes!