Created Sound scriptable objects, AudioManager class, & test sounds
AudioManager class, which is attached to an AudioManager gameobject, creates YarnCommands "PlaySound" and "StopSound" which allow sounds to be started and stopped from Yarn (i.e. rain sfx). It also has method playCharacterSound that speaks with the DialogueManager to find the current speaker, gets the speaker's Character Manager's active sound, and then plays it. This method is called by the Yarn Dialogue UI script's On Line Start method. Also added to the CharacterManager class an array of Sound scriptable objects. Each character will have at least one sound, but certain characters may have several (like Daniella having a normal and distressed sound). Thus the CharacterManager class also adds a new Yarn command, SetSound, which can allow the character's active sound to be changed between the Sound objects in their array via Yarn. The first Sound in every character's array is automatically set to be the active sound on Awake, so if a character only has one sound, they never need a SetSound command call.

In addition to all that, there are now Sound scriptable objects. These contain an audio clip, a name, and then a couple parameters that directly map onto an Audio Source's variables (Loop, Play on Awake, Volume). This way we can easily make an object for each sound in the game. The AudioManager instantiates children gameobjects at the top of the game that each have one audio source component according to the set array of sounds you give the AudioManager in the editor. If you want a sound (including music) to be played in a scene, you have to add it to the AudioManager's array of sounds.

That's about it fam. Other than that, updated the character prefab for Daniella with a test character sound and created a filler background music sound that is currently a Lizzo x Stardew Valley mashup.