Changed OptionsTextDisplayHandler > OptionsHandler, made TitleMenu scene
OptionTextDisplayHandler > OptionsHandler. Still updates the color of the options bubbles but makes them selectable with the keyboard as well. For each scene, the buttons navigation should be set to Automatic.

Also started the TitleMenu scene. Added the Start Menu, Options, and Act Select screens. No implementation yet, just the gameobjects and components.