Added the car & night scenes, updated prefabs, added some assets
All of the car and night scenes now exist and are hooked up to the dialogue system/canvas. They have empty audio managers as well as any art that is ready for them. The car scenes all have the proper Daniella and Mish characters according to who is driving. Daniella and Mish's prefabs are also updated with Far and Close bubble positions for the side shot.