Finished StartMenu, added audio, updated Sound.cs, updated options
Imported Plum's audio assets from beta sprint 1. Adjusted Sound.cs to include an enum for the type of audio (music, character, sfx) so that the volume settings can adjust the sound object's volume accordingly. Added the rest of the pop-ups to the start menu and fixed button navigation for them all. Linked a (currently empty) hover sfx to play whenever a button is selected (including the options buttons in the dialogue system). Adjusted the button navigation system in OptionsHandler. Added ExitGame.cs and ToggleOptionText.cs which allow for a button to exit the game .exe and for toggle UI components' label text to change in accordance to their bool value. Created Tooltip.cs which changes the tooltip text on the StartMenu according to what button is selected. Updated Daniella's character prefabs to include her internal sound.