Made Options menu work, Dialogue Manager & Sound.cs now refer 2 options
The options menu now changes player prefs and saves them in between file runs. The Dialogue Manager now grabs the text speed and adjusts according to player prefs, and the Sound files upon initialization set their volume according to the player prefs as well. The player can now also delete their save data from the options menu. Fullscreen toggling also should work (but needs to be tested in a .exe file). The auto advance option can be toggled and it is saved, but the auto advance feature itself is not yet implemented.