Made ForgivenessTracker (loads & saves forgive), updated audio settings
Made the Forgiveness Tracker and added it to the DialogueManager prefab and connected it to each car and motel scene. At the start of dialogue, it grabs the forgiveness value (saved in player prefs, default is 0) and sets the value in Yarn's memory to that value. At the end of a dialogue, it gets the forgiveness value from Yarn's memory and saves it to player prefs.

The audio manager can now update sound volumes (so that the audio on the start menu changes as you change it). Whenever you change the sfx or character sounds, it will play a sound at the new volume level as confirmation.