Made the credits scroll, added some special thanks, added logos
The credits now scroll on the title menu. They also appear at the end of the game (Act 3 Chapter 4) and scroll there as well. On the title menu, the full logo gently bobs up and down in the middle of the start menu. The .app and .exe logo has also been imported and added to the ProjectSettings.